new modules added and updated

+++ app/Http/Controllers/API/EnrollmentController.php
@@ -165,6 +165,6 @@ class EnrollmentController extends Controller
     }
     public function verify($id)
     {
-        return Enrollment::isExist($id);
+        return (int)Enrollment::isExist($id);
     }
 }
